+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Fix default object sprites retaining dimensions and transparency of the previous sprite.
 - Fix connections not being deleted when the map name text box is cleared.
 - Fix the map border not updating when a tileset is changed.
+- Stop the Tileset Editor from scrolling to the initially selected metatile when saving.
 
 ## [5.1.1] - 2023-02-20
 ### Added